Stop hiding JNA's address family constants
JNA's IPHlpAPI already declares AF_UNSPEC, AF_INET and AF_INET6 as int, so
redeclaring them as short in the interface extending it hid all three behind a
narrower type. Anything reading IPHlpAPI.AF_INET would silently get a short
depending only on which of the two interfaces it imported.

They are removed in favour of the inherited constants, narrowed at the one call
site whose parameter is an ADDRESS_FAMILY.

The layout test gains the two offsets the existing fixture could not reach:
both unions carry both arms whatever their family field says, so reading them
directly is what pins OFFSET_ROUTE_DEST_IPV6 and OFFSET_ROUTE_NEXTHOP_IPV4,
which the backends use for the opposite family from the one this row declares.

+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsNetworkParamsJNA.java
@@ -101,7 +101,7 @@ final class WindowsNetworkParamsJNA extends WindowsNetworkParams {
     @Override
     protected List<RouteRow> queryRouteRows() {
         try (CloseablePointerByReference tableRef = new CloseablePointerByReference()) {
-            int ret = IPHlpAPI.INSTANCE.GetIpForwardTable2(IPHlpAPI.AF_UNSPEC, tableRef);
+            int ret = IPHlpAPI.INSTANCE.GetIpForwardTable2((short) IPHlpAPI.AF_UNSPEC, tableRef);
             if (ret != WinError.NO_ERROR) {
                 LOG.error("Failed to get the IP forward table. Error code: {}", ret);
                 return new ArrayList<>();
